+++ snxcore/src/platform/linux/xfrm.rs
@@ -391,7 +391,7 @@ impl XfrmConfigurator {
                     .tunnel_params
                     .ignore_search_domains
                     .iter()
-                    .any(|d| d.to_lowercase() == s.to_lowercase())
+                    .any(|d| d.to_lowercase() == s.trim_matches('~').to_lowercase())
             })
             .cloned()
             .collect::<Vec<_>>();