+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## v3.0.7 (unreleased)
-- Added `password-factor=N` option to determine which authentication factor is the password.
+## v3.1.0 (unreleased)
+- Added `password-factor=N` option to determine which authentication factor is the password. Default is 1 (first).
+- Added `set-routing-domains=true|false` option to treat the received search domains as [routing domains](https://systemd.io/RESOLVED-VPNS/).
 
 ## v3.0.6 (2025-02-16)
 - Improved authentication prompts by displaying a header retrieved from VPN server